Update Stringtable for May Update
Updated Spanish translations for new strings, fixed some artifacts that I missed in my previous debugging. 

When checking this merge I recommend checking for feedback such as "added spanish translation" injected randomly into the code. I hope to do more work in bringing languages to this project but I need to work on my ChatGPT instance a little more so that it doesn't try to fight me so much before I can proceed, I had to constantly keep the AI on track while processing the file piecemeal which was not a fun task.